pyproject.toml
[tool.poetry]
name = "ese200"
version = "0.1.0"
description = "Dynamical Systems Labs for ESE2000"
authors = ["Damian Owerko <Damowerko@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.*"
numpy = "^1.24.2"
scipy = "^1.10.1"
torch = "^2.0.0"
cvxpy = "^1.3.1"
matplotlib = "^3.7.1"
tqdm = "^4.65.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"